+++ app/helpers/admin/team_members_helper.rb
@@ -1,0 +1,2 @@
+module Admin::TeamMembersHelper
+end