+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 8.3.0
 
 ## 8.2.2 (September 28, 2015)
+- **Admin Changes**
+  - Team members can be modified from the admin dashboard
 
 ### Enhancements
 - **General Changes**